requirements.txt
#  pip freeze > requirements.txt
aioquic==1.2.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
apacheconfig==0.3.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cmake==4.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
Cython==3.2.4
defusedxml==0.7.1
dnspython==2.8.0
easydict==1.13
faiss-cpu==1.12.0
fastapi==0.119.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.10.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.54.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.10
ImageIO==2.37.2
insightface @ file:///D:/WUP_PROJECT/python/insightface-0.7.3-cp312-cp312-win_amd64.whl#sha256=4e58a504433ba5a500d48328689e7d6c69873165653ded7553ce804beb8723db
invoke==2.2.1
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaitaistruct==0.10
kiwisolver==1.4.9
lazy_loader==0.4
ldap3==2.9.1
loguru==0.7.3
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.8
mitmproxy==12.1.2
mitmproxy-windows==0.12.7
mitmproxy_rs==0.12.7
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.4
onnx==1.20.1
opencv-python-headless==4.13.0.90
packaging==25.0
paramiko==4.0.0
passlib==1.7.4
pdfkit==1.0.0
pillow==12.0.0
ply==3.11
prettytable==3.17.0
priority==2.0.0
protobuf==6.33.4
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydivert==2.1.0
pylsqpack==0.3.22
pymongo==4.15.3
PyMuPDF==1.26.1
PyNaCl==1.6.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-whois==0.9.6
PyYAML==6.0.2
redis==7.1.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
service-identity==24.2.0
setuptools==80.9.0
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.48.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.1.14
tokenizers==0.22.1
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
urwid==2.6.16
uvicorn==0.37.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
zstandard==0.24.0
